+++ src/_data/index.js
@@ -1,0 +1,211 @@
+export const properties = [
+  {
+    id: 1,
+    title: "Property 1",
+    address: "123 Main St",
+    description: "Description for property 1",
+    price: "$100000",
+    images: [
+      "https://via.placeholder.com/300x150",
+      "https://via.placeholder.com/300x150",
+      "https://via.placeholder.com/300x150",
+      "https://via.placeholder.com/300x150",
+    ],
+    rooms: 3,
+    bathrooms: 2,
+    garage: 1,
+    surface: "150",
+    type: "casa",
+    operation: "venta",
+    reserved: false,
+    age: 5,
+    createdAt: "2023-01-01",
+  },
+  {
+    id: 2,
+    title: "Property 2",
+    address: "456 Elm St",
+    description: "Description for property 2",
+    price: "$200000",
+    images: ["https://via.placeholder.com/300x150"],
+    rooms: 4,
+    bathrooms: 3,
+    garage: 2,
+    surface: "200",
+    type: "departamento",
+    operation: "alquiler",
+    reserved: true,
+    age: 10,
+    createdAt: "2023-02-01",
+  },
+  {
+    id: 3,
+    title: "Property 3",
+    address: "789 Oak St",
+    description: "Description for property 3",
+    price: "$300000",
+    images: ["https://via.placeholder.com/300"],
+    rooms: 2,
+    bathrooms: 1,
+    garage: 1,
+    surface: "120",
+    type: "pozo",
+    operation: "venta",
+    reserved: false,
+    age: 1,
+    createdAt: "2023-03-01",
+  },
+  {
+    id: 4,
+    title: "Property 4",
+    address: "101 Pine St",
+    description: "Description for property 4",
+    price: "$400000",
+    images: ["https://via.placeholder.com/150"],
+    rooms: 5,
+    bathrooms: 4,
+    garage: 3,
+    surface: "250",
+    type: "casa",
+    operation: "alquiler",
+    reserved: true,
+    age: 15,
+    createdAt: "2023-04-01",
+  },
+  {
+    id: 5,
+    title: "Property 5",
+    address: "202 Maple St",
+    description: "Description for property 5",
+    price: "$500000",
+    images: ["https://via.placeholder.com/100x150"],
+    rooms: 6,
+    bathrooms: 5,
+    garage: 4,
+    surface: "300",
+    type: "departamento",
+    operation: "venta",
+    reserved: false,
+    age: 20,
+    createdAt: "2023-05-01",
+  },
+  {
+    id: 6,
+    title: "Property 6",
+    address: "303 Birch St",
+    description: "Description for property 6",
+    price: "$600000",
+    images: ["https://via.placeholder.com/300x250"],
+    rooms: 7,
+    bathrooms: 6,
+    garage: 5,
+    surface: "350",
+    type: "departamento",
+    operation: "alquiler",
+    reserved: true,
+    age: 25,
+    createdAt: "2023-06-01",
+  },
+  {
+    id: 7,
+    title: "Property 7",
+    address: "404 Cedar St",
+    description: "Description for property 7",
+    price: "$700000",
+    images: ["https://via.placeholder.com/200x150"],
+    rooms: 8,
+    bathrooms: 7,
+    garage: 6,
+    surface: "400",
+    type: "casa",
+    operation: "venta",
+    reserved: false,
+    age: 30,
+    createdAt: "2023-07-01",
+  },
+  {
+    id: 8,
+    title: "Property 8",
+    address: "505 Walnut St",
+    description: "Description for property 8",
+    price: "$800000",
+    images: ["https://via.placeholder.com/250x200"],
+    rooms: 9,
+    bathrooms: 8,
+    garage: 7,
+    surface: "450",
+    type: "casa",
+    operation: "alquiler",
+    reserved: true,
+    age: 35,
+    createdAt: "2023-08-01",
+  },
+  {
+    id: 9,
+    title: "Property 9",
+    address: "606 Chestnut St",
+    description: "Description for property 9",
+    price: "$900000",
+    images: ["https://via.placeholder.com/350x150"],
+    rooms: 10,
+    bathrooms: 9,
+    garage: 8,
+    surface: "500",
+    type: "departamento",
+    operation: "venta",
+    reserved: false,
+    age: 40,
+    createdAt: "2023-09-01",
+  },
+  {
+    id: 10,
+    title: "Property 10",
+    address: "707 Spruce St",
+    description: "Description for property 10",
+    price: "$1000000",
+    images: ["https://via.placeholder.com/300x200"],
+    rooms: 11,
+    bathrooms: 10,
+    garage: 9,
+    surface: "550",
+    type: "pozo",
+    operation: "alquiler",
+    reserved: true,
+    age: 45,
+    createdAt: "2023-10-01",
+  },
+  {
+    id: 11,
+    title: "Property 11",
+    address: "808 Fir St",
+    description: "Description for property 11",
+    price: "$1100000",
+    images: ["https://via.placeholder.com/400x300"],
+    rooms: 12,
+    bathrooms: 11,
+    garage: 10,
+    surface: "600",
+    type: "casa",
+    operation: "venta",
+    reserved: false,
+    age: 50,
+    createdAt: "2023-11-01",
+  },
+  {
+    id: 12,
+    title: "Property 12",
+    address: "909 Redwood St",
+    description: "Description for property 12",
+    price: "$1200000",
+    images: ["https://via.placeholder.com/150x100"],
+    rooms: 13,
+    bathrooms: 12,
+    garage: 11,
+    surface: "650",
+    type: "departamento",
+    operation: "alquiler",
+    reserved: true,
+    age: 55,
+    createdAt: "2023-12-01",
+  },
+];
