+++ src/pages/Detail/index.jsx
@@ -7,9 +7,10 @@ import {
   FaBed,
   FaCalendarAlt,
 } from "react-icons/fa";
-import { properties } from "../Propiedades";
+import { properties } from "../../_data";
 import { useParams, useNavigate } from "react-router-dom";
 import ImageGallery from "react-image-gallery";
+import { formatPrice } from "../../utils";
 import "./styles.css";
 
 const Detail = () => {
@@ -38,10 +39,15 @@ const Detail = () => {
         <Card.Body>
           <Row className="mb-2">
             <Col>
-              <Badge bg="light" text="dark" className="me-2">
-                Venta
+              <Badge
+                bg="light"
+                text="dark"
+                className="me-2"
+                style={{ textTransform: "capitalize" }}
+              >
+                {property.operation}
               </Badge>
-              <Badge bg="primary">Reservada</Badge>
+              {property.reserved && <Badge bg="primary">Reservada</Badge>}
             </Col>
           </Row>
 
@@ -49,38 +55,37 @@ const Detail = () => {
             className="mb-2"
             style={{ fontSize: "24px", fontWeight: "bold" }}
           >
-            45.000 USD
+            {formatPrice(property.price)} USD
           </Card.Title>
 
-          <Card.Text className="text-muted">Expensas: 2.000 ARS</Card.Text>
+          {property.operation === "alquiler" && (
+            <Card.Text className="text-muted">Expensas: 90.000 ARS</Card.Text>
+          )}
 
           <Row className="mb-3">
             <Col xs="auto">
               <FaRulerCombined className="me-2" />
-              <strong>54.79</strong> m² totales
+              <strong>{property.surface}</strong> m² totales
             </Col>
             <Col xs="auto">
               <FaRulerCombined className="me-2" />
-              <strong>54.79</strong> m² cubiertos
+              <strong>{property.surface}</strong> m² cubiertos
             </Col>
             <Col xs="auto">
               <FaBuilding className="me-2" />
-              <strong>3</strong> ambientes
+              <strong>{property.rooms}</strong> ambientes
             </Col>
             <Col xs="auto">
               <FaBath className="me-2" />
-              <strong>1</strong> baño
+              <strong>{property.bathrooms}</strong> baño
             </Col>
-          </Row>
-
-          <Row>
             <Col xs="auto">
               <FaBed className="me-2" />
-              <strong>2</strong> dormitorios
+              <strong>{property.rooms}</strong> dormitorios
             </Col>
             <Col xs="auto">
               <FaCalendarAlt className="me-2" />
-              <strong>22</strong> años antigüedad
+              <strong>{property.age}</strong> años antigüedad
             </Col>
           </Row>
         </Card.Body>
