+++ src/pages/Propiedades/Filters/Chips/index.jsx
@@ -2,12 +2,12 @@ import React, { useState } from "react";
 import { Badge, Container, Stack } from "react-bootstrap";
 
 const FilterChips = ({ onFilterChange, filters }) => {
-  const [selected, setSelected] = useState(0);
+  const [selected, setSelected] = useState(filters.rooms || 0);
 
   const handleSelect = (index) => {
     setSelected(index);
-    const bedrooms = index === 0 ? 0 : index;
-    onFilterChange({ ...filters, bedrooms });
+    const rooms = index === 0 ? 0 : index;
+    onFilterChange({ ...filters, rooms });
   };
 
   const badges = [
@@ -19,24 +19,22 @@ const FilterChips = ({ onFilterChange, filters }) => {
   ];
 
   return (
-    <Container className="m-2">
-      <Stack direction="horizontal" gap={2} className="flex-wrap">
-        {badges.map((badge, index) => (
-          <Badge
-            key={index}
-            bg="white"
-            text="dark"
-            className={`p-2 border ${
-              selected === index ? "border-3 border-dark" : ""
-            }`}
-            onClick={() => handleSelect(index)}
-            style={{ cursor: "pointer", fontSize: "0.9rem" }}
-          >
-            {badge}
-          </Badge>
-        ))}
-      </Stack>
-    </Container>
+    <Stack direction="horizontal" gap={2} className="flex-wrap">
+      {badges.map((badge, index) => (
+        <Badge
+          key={index}
+          bg="white"
+          text="dark"
+          className={`p-2 border ${
+            selected === index ? "border-3 border-dark" : ""
+          }`}
+          onClick={() => handleSelect(index)}
+          style={{ cursor: "pointer", fontSize: "0.9rem" }}
+        >
+          {badge}
+        </Badge>
+      ))}
+    </Stack>
   );
 };
