+++ src/store/index.jsx
@@ -1,8 +1,15 @@
 import { create } from "zustand";
 
 const useFilterStore = create((set) => ({
-  filters: {},
-  setFilters: (newFilters) => set({ filters: newFilters }),
+  filters: {
+    address: "",
+    type: "",
+    minPrice: "",
+    maxPrice: "",
+  },
+  sortKey: { key: "", order: 1 },
+  setFilters: (filters) => set({ filters }),
+  setSortKey: (sortKey) => set({ sortKey }),
 }));
 
 export default useFilterStore;
